# src/django-backend/api/routers/projects.py
from typing import Any

from django.http import HttpRequest
from ninja import Query, Router

from api.auth.security import auth
from api.rate_limit import check_rate_limit
from api.routers._helpers import get_optional_user
from api.schemas.errors import Error
from api.schemas.project import (
    CategoryResponse,
    DiscoverProjectResponse,
    ProjectListItemResponse,
    ProjectListResponse,
    ProjectResponse,
    WinnerProjectResponse,
)
from api.schemas.project_report import ProjectReportCreate, ProjectReportResponse
from apps.projects.models import Project, ProjectReport, ProjectStatus
from services import HANDLERS, REPO
from services.project.exceptions import ProjectNotFoundError
from services.project.handler_interface import ReportProjectInput

router = Router()


@router.get(
    "/categories",
    response={200: list[CategoryResponse]},
    tags=["Projects"],
)
def list_categories(request: HttpRequest) -> list[dict[str, Any]]:
    return [
        {
            "id": c.id,
            "name": c.name,
            "slug": c.slug,
            "project_count": c.project_count,
        }
        for c in REPO.project.list_categories()
    ]


@router.get(
    "/featured",
    response={200: list[DiscoverProjectResponse]},
    tags=["Projects"],
)
def list_featured(request: HttpRequest) -> list[DiscoverProjectResponse]:
    return [
        DiscoverProjectResponse.from_discover_item(item)
        for item in REPO.project.list_featured()
    ]


@router.get(
    "/new-arrivals",
    response={200: list[DiscoverProjectResponse]},
    tags=["Projects"],
)
def list_new_arrivals(request: HttpRequest) -> list[DiscoverProjectResponse]:
    return [
        DiscoverProjectResponse.from_discover_item(item)
        for item in REPO.project.list_new_arrivals()
    ]


@router.get(
    "/recent-tipoffs",
    response={200: list[DiscoverProjectResponse]},
    tags=["Projects"],
)
def list_recent_tipoffs(request: HttpRequest) -> list[DiscoverProjectResponse]:
    return [
        DiscoverProjectResponse.from_discover_item(item)
        for item in REPO.project.list_recent_tipoffs()
    ]


@router.get(
    "/winners",
    response={200: list[WinnerProjectResponse]},
    tags=["Projects"],
)
def list_winners(request: HttpRequest) -> list[WinnerProjectResponse]:
    return [
        WinnerProjectResponse(
            id=w.project.id,
            slug=w.project.slug,
            title=w.project.title,
            tagline=w.project.tagline,
            icon_url=w.icon_url,
            hero_banner_url=w.hero_banner_url,
            in_use_image_url=w.in_use_image_url,
            competition_name=w.competition_name,
            competition_slug=w.competition_slug,
            competition_submission_deadline=w.competition_submission_deadline.isoformat()
            if w.competition_submission_deadline
            else "",
        )
        for w in REPO.project.list_winners()
    ]


@router.get(
    "/most-discussed",
    response={200: list[DiscoverProjectResponse]},
    tags=["Projects"],
)
def list_most_discussed(request: HttpRequest) -> list[DiscoverProjectResponse]:
    return [
        DiscoverProjectResponse.from_discover_item(item)
        for item in REPO.project.list_most_discussed()
    ]


@router.get(
    "/by-category/{slug}",
    response={200: list[DiscoverProjectResponse], 404: Error},
    tags=["Projects"],
)
def list_by_category(
    request: HttpRequest,
    slug: str,
    sort: str = Query("newest"),
) -> list[DiscoverProjectResponse] | tuple[int, dict[str, str]]:
    try:
        items = REPO.project.list_by_category(slug, sort)
    except ProjectNotFoundError:
        return 404, {"detail": "Category not found"}
    return [DiscoverProjectResponse.from_discover_item(item) for item in items]


@router.get("", response={200: ProjectListResponse, 400: Error}, tags=["Projects"])
def list_projects(
    request: HttpRequest,
    tags: list[str] | None = Query(None),
    tech_stack: list[str] | None = Query(None),
    sort_by: str = Query("created_at"),
    sort_order: str = Query("desc"),
    search: str | None = Query(None),
    page: int = Query(1),
    per_page: int = Query(20),
) -> dict[str, Any] | tuple[int, dict[str, str]]:
    try:
        result = REPO.project.list_approved(
            tags=tags,
            tech_stack=tech_stack,
            search=search,
            sort_by=sort_by,
            sort_order=sort_order,
            page=page,
            per_page=per_page,
        )
    except ValueError as e:
        return 400, {"detail": str(e)}
    return {
        "projects": [
            ProjectListItemResponse.from_list_item(p) for p in result.projects
        ],
        "total": result.total,
        "page": result.page,
        "per_page": result.per_page,
        "pages": result.pages,
        "pending_projects_count": REPO.project.count_pending(),
    }


@router.get(
    "/{identifier}",
    response={200: ProjectResponse, 404: Error},
    tags=["Projects"],
)
def get_project(
    request: HttpRequest,
    identifier: str,
) -> Project | tuple[int, dict[str, str]]:
    try:
        project = REPO.project.get_by_identifier(identifier)
    except ProjectNotFoundError:
        return 404, {"detail": "Project not found"}

    user = get_optional_user(request)
    user_id = user.id if user else None
    project.is_followed = REPO.follows.is_followed(user_id, project)

    if project.status == ProjectStatus.APPROVED:
        return project

    if user and user.is_superuser:
        return project
    if REPO.project.user_can_edit(project.id, user_id):
        return project

    return 404, {"detail": "Project not found"}


@router.post(
    "/{identifier}/reports",
    response={201: ProjectReportResponse, 401: Error, 404: Error, 429: Error},
    auth=auth,
    tags=["Projects"],
)
def report_project(
    request: HttpRequest,
    identifier: str,
    payload: ProjectReportCreate,
) -> Any:
    # Signed-in users only: an anonymous form let anyone send email to a maker
    # in Naglasúpan's name, and reports marked as spam cost the domain its
    # sending reputation. The limit is per account, so users behind one NAT
    # don't share it.
    rate_limit_response = check_rate_limit(
        request, "report_project", "5/h", key=str(request.auth.id)
    )
    if rate_limit_response is not None:
        return rate_limit_response

    try:
        project = REPO.project.get_by_identifier(identifier)
    except ProjectNotFoundError:
        return 404, {"detail": "Project not found"}

    try:
        report: ProjectReport = HANDLERS.project.report(
            ReportProjectInput(
                project_id=project.id,
                reason=payload.reason,
                details=payload.details,
                contact_email=payload.contact_email,
                reporter_id=request.auth.id,
            )
        )
    except ProjectNotFoundError:
        return 404, {"detail": "Project not found"}
    return 201, report