# src/web-ui/backend-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Project Showcase API",
    "version": "1.0.0",
    "description": "API for developer project showcasing platform"
  },
  "paths": {
...
    "/api/projects/by-category/{slug}": {
      "get": {
        "operationId": "api_routers_projects_list_by_category",
        "summary": "List By Category",
        "parameters": [
          {
            "in": "path",
            "name": "slug",
            "schema": {
              "title": "Slug",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "default": "newest",
              "title": "Sort",
              "type": "string"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DiscoverProjectResponse"
                  },
                  "title": "Response",
                  "type": "array"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": [
          "Projects"
        ]
      }
    },
    "/api/projects": {
      "get": {
        "operationId": "api_routers_projects_list_projects",
        "summary": "List Projects",
        "parameters": [
          {
            "in": "query",
            "name": "tags",
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tags"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "tech_stack",
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tech Stack"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "sort_by",
            "schema": {
              "default": "created_at",
              "title": "Sort By",
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "sort_order",
            "schema": {
              "default": "desc",
              "title": "Sort Order",
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "default": 1,
              "title": "Page",
              "type": "integer"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "per_page",
            "schema": {
              "default": 20,
              "title": "Per Page",
              "type": "integer"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectListResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": [
          "Projects"
        ]
      }
    },
    "/api/projects/{identifier}": {
      "get": {
        "operationId": "api_routers_projects_get_project",
        "summary": "Get Project",
        "parameters": [
          {
            "in": "path",
            "name": "identifier",
            "schema": {
              "title": "Identifier",
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": [
          "Projects"
        ]
      }
    },
    "/api/projects/{identifier}/reports": {
      "post": {
        "operationId": "api_routers_projects_report_project",
        "summary": "Report Project",
        "parameters": [
          {
            "in": "path",
            "name": "identifier",
            "schema": {
              "title": "Identifier",
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectReportResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Too Many Requests",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": [
          "Projects"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectReportCreate"
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "JWTAuth": []
          }
        ]
      }
    },
...
    "/api/my/projects": {
      "get": {
        "operationId": "api_routers_my_projects_list_my_projects",
        "summary": "List My Projects",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProjectResponse"
                  },
                  "title": "Response",
                  "type": "array"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": [
          "My Projects"
        ],
        "security": [
          {
            "JWTAuth": []
          }
        ]
      },
      "post": {
        "operationId": "api_routers_my_projects_create_project",
        "summary": "Create Project",
        "parameters": [],
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": [
          "My Projects"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "JWTAuth": []
          }
        ]
      }
    },
...
    "/api/my/projects/{project_id}": {
      "get": {
        "operationId": "api_routers_my_projects_get_my_project",
        "summary": "Get My Project",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "schema": {
              "title": "Project Id",
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": [
          "My Projects"
        ],
        "security": [
          {
            "JWTAuth": []
          }
        ]
      },
      "put": {
        "operationId": "api_routers_my_projects_update_project",
        "summary": "Update Project",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "schema": {
              "title": "Project Id",
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": [
          "My Projects"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectCreate"
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "JWTAuth": []
          }
        ]
      },
      "delete": {
        "operationId": "api_routers_my_projects_delete_project",
        "summary": "Delete Project",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "schema": {
              "title": "Project Id",
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": [
          "My Projects"
        ],
        "security": [
          {
            "JWTAuth": []
          }
        ]
      }
    },
...
    "/api/my/reviews/projects/{project_id}": {
      "get": {
        "operationId": "api_routers_my_review_get_review_project",
        "summary": "Get Review Project",
        "parameters": [
          {
            "in": "path",
            "name": "project_id",
            "schema": {
              "title": "Project Id",
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewProjectDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "description": "Get project details for a reviewer.\n\nReturns the project if the user is assigned as a reviewer to any\ncompetition that contains this project. Returns 404 otherwise.",
        "tags": [
          "My Review"
        ],
        "security": [
          {
            "JWTAuth": []
          }
        ]
      }
    },
...
    "/api/users/{user_id}/projects": {
      "get": {
        "operationId": "api_routers_users_list_public_projects",
        "summary": "List Public Projects",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/UserProjectResponse"
                  },
                  "title": "Response",
                  "type": "array"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Error": {
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "Error",
        "type": "object"
      },
...
      "DiscoverProjectResponse": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "tagline": {
            "title": "Tagline",
            "type": "string"
          },
          "icon_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon Url"
          },
          "hero_banner_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hero Banner Url"
          },
          "in_use_image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "In Use Image Url"
          },
          "category_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Name"
          },
          "category_slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Slug"
          },
          "discussion_count": {
            "default": 0,
            "title": "Discussion Count",
            "type": "integer"
          },
          "won_competitions": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/WonCompetitionInfo"
            },
            "title": "Won Competitions",
            "type": "array"
          },
          "is_community_tipoff": {
            "default": false,
            "title": "Is Community Tipoff",
            "type": "boolean"
          }
        },
        "required": [
          "id",
          "slug",
          "title",
          "tagline"
        ],
        "title": "DiscoverProjectResponse",
        "type": "object"
      },
...
      "ProjectListResponse": {
        "properties": {
          "projects": {
            "items": {
              "$ref": "#/components/schemas/ProjectListItemResponse"
            },
            "title": "Projects",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "per_page": {
            "title": "Per Page",
            "type": "integer"
          },
          "pages": {
            "title": "Pages",
            "type": "integer"
          },
          "pending_projects_count": {
            "title": "Pending Projects Count",
            "type": "integer"
          }
        },
        "required": [
          "projects",
          "total",
          "page",
          "per_page",
          "pages",
          "pending_projects_count"
        ],
        "title": "ProjectListResponse",
        "type": "object"
      },
...
      "ProjectResponse": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "tagline": {
            "title": "Tagline",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "long_description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Long Description"
          },
          "website_url": {
            "title": "Website Url",
            "type": "string"
          },
          "github_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Github Url"
          },
          "demo_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Demo Url"
          },
          "tech_stack": {
            "items": {
              "type": "string"
            },
            "title": "Tech Stack",
            "type": "array"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "approved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approved At"
          },
          "published_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "owner": {
            "$ref": "#/components/schemas/PublicUserProfile"
          },
          "creator": {
            "$ref": "#/components/schemas/PublicUserProfile"
          },
          "contributors": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ContributorSummary"
            },
            "title": "Contributors",
            "type": "array"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/TagWithCategoryResponse"
            },
            "title": "Tags",
            "type": "array"
          },
          "images": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/ProjectImageResponse"
            },
            "title": "Images",
            "type": "array"
          },
          "won_competitions": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/WonCompetitionInfo"
            },
            "title": "Won Competitions",
            "type": "array"
          },
          "is_community_tipoff": {
            "default": false,
            "title": "Is Community Tipoff",
            "type": "boolean"
          },
          "is_followed": {
            "default": false,
            "title": "Is Followed",
            "type": "boolean"
          },
          "competition_standing": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CompetitionStandingResponse"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "slug",
          "title",
          "tagline",
          "description",
          "long_description",
          "website_url",
          "github_url",
          "demo_url",
          "tech_stack",
          "status",
          "created_at",
          "approved_at",
          "published_at",
          "owner",
          "creator",
          "tags"
        ],
        "title": "ProjectResponse",
        "type": "object"
      },
...
      "ProjectReportResponse": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "reason": {
            "enum": [
              "site_down",
              "something_broken",
              "wrong_link",
              "other"
            ],
            "title": "Reason",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "reason",
          "created_at"
        ],
        "title": "ProjectReportResponse",
        "type": "object"
      },
      "ProjectReportCreate": {
        "properties": {
          "reason": {
            "enum": [
              "site_down",
              "something_broken",
              "wrong_link",
              "other"
            ],
            "title": "Reason",
            "type": "string"
          },
          "details": {
            "default": "",
            "maxLength": 2000,
            "title": "Details",
            "type": "string"
          },
          "contact_email": {
            "default": "",
            "maxLength": 254,
            "title": "Contact Email",
            "type": "string"
          }
        },
        "required": [
          "reason"
        ],
        "title": "ProjectReportCreate",
        "type": "object"
      },
...
      "UserProjectResponse": {
        "description": "One entry in a user's public project list.\n\n`role` is the contributor row's role. `full_edit` is deliberately absent:\nit is an editing permission, not a public fact.",
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "slug": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Slug"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "tagline": {
            "title": "Tagline",
            "type": "string"
          },
          "category_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category Name"
          },
          "main_image_thumb_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Main Image Thumb Url"
          },
          "role": {
            "enum": [
              "owner",
              "tipster"
            ],
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "id",
          "slug",
          "title",
          "tagline",
          "category_name",
          "main_image_thumb_url",
          "role"
        ],
        "title": "UserProjectResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "JWTAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": []
}